+++ CsvSorter.Tests/StreamReaderExtensionsTests.cs
@@ -1,27 +1,28 @@
-﻿using NUnit.Framework;
-
-namespace CsvSorter.Tests;
+﻿namespace CsvSorter.Tests;
 
 public class StreamReaderExtensionsTests
 {
-
     [Test]
-    public void OrderByExtensionReturnsCsvSorterWithCorrectGenericType()
+    public void Ensure_GetCsvSorterBy_Extension_Returns_CsvSorter_With_Correct_Generic_Type()
     {
-        var sorter = new StreamReader(new MemoryStream())
-            .OrderBy<DateTime>(0);
+        using var sr = new StreamReader(new MemoryStream());
+        
+        var sorterByFieldIndex = sr
+            .GetCsvSorter<DateTime>(0);
 
-        Assert.NotNull(sorter);
-        Assert.That(sorter.GetType() == typeof(CsvSorter<DateTime>));
-    }
-
-    [Test]
-    public void OrderByDescendingExtensionReturnsCsvSorterWithCorrectGenericType()
-    {
-        var sorter = new StreamReader(new MemoryStream())
-            .OrderByDescending<string>(0);
-
-        Assert.NotNull(sorter);
-        Assert.That(sorter.GetType() == typeof(CsvSorter<string>));
+        var sorterByFieldName = sr
+            .GetCsvSorter<bool>("any");
+        
+        Assert.Multiple(() =>
+        {
+            Assert.That(sorterByFieldIndex, Is.Not.Null);
+            Assert.That(sorterByFieldName, Is.Not.Null);
+        });
+        
+        Assert.Multiple(() =>
+        {
+            Assert.That(sorterByFieldIndex.GetType(), Is.EqualTo(typeof(CsvSorter<DateTime>)));
+            Assert.That(sorterByFieldName.GetType(), Is.EqualTo(typeof(CsvSorter<bool>)));
+        });
     }
 }