+++ CsvSorter/Entities/CsvSorterIndex.cs
@@ -1,8 +1,8 @@
-﻿namespace CsvSorter.Entities;
+﻿namespace CsvSorter;
 
-public class CsvSorterIndex<T> where T: IComparable
+public class CsvSorterIndex<T> where T: IComparable<T>
 {
-    public T Value { get; set; }
+    public T? Value { get; set; }
     public long Offset { get; set; }
     public int Length { get; set; }
 }