+++ CsvSorter/Entities/SortDirection.cs
@@ -1,0 +1,7 @@
+﻿namespace CsvSorter;
+
+public enum SortDirection
+{
+    Ascending,
+    Descending
+}