+++ CsvSorter/Providers/MemoryIndexProvider.cs
@@ -1,29 +1,30 @@
-﻿using CsvSorter.Entities;
+﻿namespace CsvSorter;
 
-namespace CsvSorter;
-
-public class MemoryIndexProvider<T> : IIndexProvider<T> where T: IComparable
+public class MemoryIndexProvider<T> : IIndexProvider<T> where T : IComparable<T>
 {
-    private readonly List<CsvSorterIndex<T>> _indexes;
-
-    public MemoryIndexProvider()
-    {
-        _indexes = new List<CsvSorterIndex<T>>();
-    }
+    private readonly List<CsvSorterIndex<T>> _indexes = [];
 
     public void Add(CsvSorterIndex<T> record)
     {
         _indexes.Add(record);
     }
 
-    public IEnumerable<CsvSorterIndex<T>> GetSorted(bool descending)
+    public IEnumerable<CsvSorterIndex<T>> GetSorted(SortDirection sortDirection)
     {
-        if (descending)
-            return _indexes
-                .OrderByDescending(x => x.Value);
+        var comparer = sortDirection == SortDirection.Ascending
+            ? Comparer<T?>.Create((x, y) =>
+            {
+                if (x == null) return y == null ? 0 : -1;
+                return y == null ? 1 : x.CompareTo(y);
+            })
+            : Comparer<T?>.Create((x, y) =>
+            {
+                if (x == null) return y == null ? 0 : 1;
+                return y == null ? -1 : y.CompareTo(x);
+            });
 
-        return _indexes
-            .OrderBy(x => x.Value);
+        foreach (var index in _indexes.OrderBy(x => x.Value, comparer))
+            yield return index;
     }
 
     public void Clear()