+++ CHANGELOG.md
@@ -1,7 +1,19 @@
+## 2.0.0
+### Features
+- Target framework is `.NET Standard 2.1` now
+- Added async index provider
+- Added `CancellationToken` to async methods
+- Added `SortDirection` options
+- Enabled nullable annotation context
+
+### Breaking changes
+- `OrderBy` and `OrderByDescending` methods are removed (Use `SortDirection` instead)
+- `StreamReader` helper name was changes to `GetCsvSorter`
+
 ## 1.1.0
 ### Features
 - Updated `CsvHelper` package
-- Updated target framework. It is `Standard2.0` now
+- Target framework is `.NET Standard 2.1` now
 
 ## 1.0.2
 ## Features